+++ CHANGELOG.md
@@ -3,6 +3,52 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](https://github.com/nextgis/nextgisweb_frontend/compare/v1.15.1...v1.16.0) (2022-08-28)
+
+
+### Features
+
+* **cesium:** add heightOffset geojson option ([caa9626](https://github.com/nextgis/nextgisweb_frontend/commit/caa9626457d28265ca169e97e006a23f9d6e452e))
+* improve geojson adapter multiselect ([30a3223](https://github.com/nextgis/nextgisweb_frontend/commit/30a32237411b60d03246bdc8efe292295ffcef44))
+* **paint:** add experimental paint 3d style ([74ddd65](https://github.com/nextgis/nextgisweb_frontend/commit/74ddd65d72fdd5539868d27da58a949ea26cd365))
+
+
+
+
+
+## [1.15.1](https://github.com/nextgis/nextgisweb_frontend/compare/v1.15.0...v1.15.1) (2022-08-02)
+
+
+### chore
+
+* **vue-ngw-map:** update dependencies ([1f2b688](https://github.com/nextgis/nextgisweb_frontend/commit/1f2b68836c1e3e367bdb1a8c2ff3652e704aae99))
+
+
+### Features
+
+* **ngw-connector:** add request transform method ([13ec477](https://github.com/nextgis/nextgisweb_frontend/commit/13ec4773aac88b6dc5880727241f3b04ef31fac0))
+
+
+### BREAKING CHANGES
+
+* **vue-ngw-map:** Vue is now required as peer dependency
+
+
+
+
+
+# [1.15.0](https://github.com/nextgis/nextgisweb_frontend/compare/v1.14.0...v1.15.0) (2022-07-27)
+
+
+### Features
+
+* **mapbox-map-adapter:** set opacity with native paint ([c227ee3](https://github.com/nextgis/nextgisweb_frontend/commit/c227ee32cb6656cdc3f2c4086b373b49e28245ca))
+* **mapbox-map-adapter:** use direct mapAdapteOptions style ([89cd703](https://github.com/nextgis/nextgisweb_frontend/commit/89cd70338f3a75748966017b5f1935f8c992bf56))
+
+
+
+
+
 # [1.14.0](https://github.com/nextgis/nextgisweb_frontend/compare/v1.13.8...v1.14.0) (2022-07-05)
 
 **Note:** Version bump only for package root