+++ packages/ol-map-adapter/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.16.0](https://github.com/nextgis/nextgis_frontend/compare/v1.15.1...v1.16.0) (2022-08-28)
+
+
+### Features
+
+* improve geojson adapter multiselect ([30a3223](https://github.com/nextgis/nextgis_frontend/commit/30a32237411b60d03246bdc8efe292295ffcef44))
+
+
+
+
+
+## [1.15.1](https://github.com/nextgis/nextgis_frontend/compare/v1.15.0...v1.15.1) (2022-08-02)
+
+
+### chore
+
+* **vue-ngw-map:** update dependencies ([1f2b688](https://github.com/nextgis/nextgis_frontend/commit/1f2b68836c1e3e367bdb1a8c2ff3652e704aae99))
+
+
+### BREAKING CHANGES
+
+* **vue-ngw-map:** Vue is now required as peer dependency
+
+
+
+
+
+# [1.15.0](https://github.com/nextgis/nextgis_frontend/compare/v1.14.0...v1.15.0) (2022-07-27)
+
+**Note:** Version bump only for package @nextgis/ol-map-adapter
+
+
+
+
+
 # [1.14.0](https://github.com/nextgis/nextgis_frontend/compare/v1.13.8...v1.14.0) (2022-07-05)
 
 **Note:** Version bump only for package @nextgis/ol-map-adapter